[compile_php7.sh] Fix PHP 8 compilation iconv error
Error: iconv does not support errno

If ext/iconv/config.m4 detects glibc's iconv implementation, it will
alwasy assume that we use it but actually we are using
/usr/local/include/iconv.h which is GNU's implementation and it needs a
"-liconv" flags to be linked correctly. This is auto done if the script
detects correctly... but it doesn't. So we forcely add the missing flag
to make the configuration test pass.

https://github.com/php/php-src/commit/66d8d0c7d7f77b608cc0b1198d6216dcde916d95

Signed-off-by: Jack Cherng <jfcherng@gmail.com>